requirements.txt
dotenv==0.9.9
mysql-connector-python==9.3.0
pip
python-dotenv==1.1.0
tk==0.1.0 
